// src/types/portfolio.ts
export type Experience = {
  company: string;
  role: string;
  period: string;
  location: string;
  project: string;
  description: string;
  technologies: string[];
};

export type CaseStudy = {
  name: string;
  image: string;
  url: string;
  category: string;
  challenge: string;
  contribution: string;
  outcome: string;
  technologies: string[];
};